# swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "description": "Authorization API for STEXS",
        "title": "Authorization API",
        "version": "1.0.0"
    },
    "host": "localhost",
    "basePath": "/auth/api/v1/",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "components": {
        "schemas": {
...
            "SuccessResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "A boolean indicator of the operation's success."
                    },
                    "message": {
                        "type": "string",
                        "description": "A human-readable message describing the result of the operation."
                    },
                    "timestamp": {
                        "type": "string",
                        "format":"ISO 8601",
                        "description": "The timestamp when the message was generated."
                    },
                    "data": {
                        "type": "object",
                        "description": "Additional relevant data returned with the response."
                    }
                },
                "example": {
                    "success": true,
                    "message": "Success message",
                    "timestamp": "2023-10-05T14:21:52.793Z",
                    "data": {}
                }
            }
        },
        "securitySchemes": {
            "Auth": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Bearer Token"
            }
        }
    },
    "paths": {
...
        "/verify": {
            "post": {
                "summary": "Verify Email Address",
                "tags": ["Verify Email"],
                "description": "Allows users to verify their email address after signing up. This endpoint is used to confirm the validity of the provided email address and activate the user's account.",
                "parameters": [
                    {
                        "name": "email",
                        "format": "email",
                        "in": "query",
                        "description": "The email address to be verified.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "The verification token sent to the user's email address.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to Sign In",
                        "headers": {
                            "Location": {
                                "description": "The URL to redirect to for sign-in with verification status.",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/mfa": {
            "get": {
                "summary": "Check MFA Status",
                "tags": ["Multi-Factor Authentication (MFA)"],
                "description": "Allows authenticated users to check the status of their Multi-Factor Authentication (MFA) flows. This endpoint provides information on whether MFA methods like email and TOTP (Time-Based One-Time Password) are enabled or disabled for the user's account.",
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "MFA Status Retrieval Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "email": {
                                            "type": "boolean",
                                            "description": "Indicates whether email-based MFA is enabled for the user's account."
                                        },
                                        "totp": {
                                            "type": "boolean",
                                            "description": "Indicates whether TOTP (Time-Based One-Time Password) MFA is enabled for the user's account."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/mfa/totp": {
            "post": {
                "summary": "Initialize Time-Based One-Time Password (TOTP) for MFA",
                "tags": ["Multi-Factor Authentication (MFA)"],
                "description": "Allows authenticated users to initialize the Time-Based One-Time Password (TOTP) method for Multi-Factor Authentication (MFA). TOTP provides an additional layer of security by generating a one-time code that users must enter during sign-in.",
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TOTP Initialization Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "secret": {
                                            "type": "string",
                                            "description": "The secret key for generating Time-Based One-Time Passwords (TOTP). Store this secret securely."
                                        },
                                        "otp_auth_uri": {
                                            "type": "string",
                                            "description": "A URI that can be used to configure authenticator apps (e.g., Google Authenticator) for TOTP. This URI includes the secret key and other necessary information."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/mfa/totp/disable": {
            "post": {
                "summary": "Disable TOTP Multi-Factor Authentication",
                "tags": ["Multi-Factor Authentication (MFA)"],
                "description": "Allows users to disable Time-Based One-Time Password (TOTP) as a Multi-Factor Authentication (MFA) method. This action will remove TOTP as an active MFA option for the user's account.",
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "description": "The confirmation code required to disable TOTP. This code is generated by the user's authenticator app."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "TOTP Multi-Factor Authentication Disabled Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/mfa/email": {
            "post": {
                "summary": "Enable Email MFA",
                "tags": ["Multi-Factor Authentication (MFA)"],
                "description": "Allows users to enable Email MFA for enhanced account security. To activate Email MFA, users must provide a confirmation code sent to their registered email address. Successful validation enables Email MFA, adding an extra layer of protection to their account.",
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "description": "The confirmation code sent to the user's registered email address."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email MFA Enable Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/mfa/email/disable": {
            "post": {
                "summary": "Disable Email MFA",
                "tags": ["Multi-Factor Authentication (MFA)"],
                "description": "Allows users to disable Email MFA for their account. To complete this process, users must provide a confirmation code sent to their registered email address. Disabling Email MFA removes the additional security layer from their account.",
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "description": "The confirmation code sent to the user's registered email address."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email MFA Disable Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/mfa/verify": {
            "post": {
                "summary": "Verify MFA",
                "tags": ["Multi-Factor Authentication (MFA)"],
                "description": "Allows users to verify the setup of MFA for their account. Users need to provide the type of MFA used and the verification code to confirm the setup of MFA for their account.",
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "description": "The type of MFA being verified (e.g., 'totp')."
                                    },
                                    "code": {
                                        "type": "string",
                                        "description": "The verification code for the specified MFA type."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "MFA Setup Verification Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/mfa/email/send-code": {
            "post": {
                "summary": "Send MFA Email Verification Code",
                "tags": ["Multi-Factor Authentication (MFA)"],
                "description": "Allows users, whether fully signed in or who have already initialized the sign-in process, to send an email verification code for MFA. Users can use this endpoint to request an email verification code.",
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "format": "JWT",
                                        "description": "A JSON Web Token (JWT) provided to users who have initialized the sign-in process."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "MFA Email Verification Code Sent Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}